chore: centralize fixtures
