+++ tests/unit/test_charm_collect_status.py
@@ -2,72 +2,18 @@
 # See LICENSE file for licensing details.
 
 import tempfile
-from unittest.mock import PropertyMock, patch
 
-import pytest
 import scenario
 from ops import ActiveStatus, BlockedStatus, WaitingStatus
 from ops.pebble import Layer, ServiceStatus
 
-from charm import AMFOperatorCharm
-from k8s_service import K8sService
 from tests.unit.certificates_helpers import (
     example_cert_and_key,
 )
-
-NRF_URL = "http://nrf:8081"
-WEBUI_URL = "sdcore-webui:9876"
-DATABASE_LIB_PATH = "charms.data_platform_libs.v0.data_interfaces.DatabaseRequires"
+from tests.unit.fixtures import AMFUnitTestFixtures
 
 
-class TestCharmCollectUnitStatus:
-    patcher_check_output = patch("charm.check_output")
-    patcher_k8s_service = patch("charm.K8sService", autospec=K8sService)
-    patcher_nrf_url = patch(
-        "charms.sdcore_nrf_k8s.v0.fiveg_nrf.NRFRequires.nrf_url", new_callable=PropertyMock
-    )
-    patcher_webui_url = patch(
-        "charms.sdcore_nms_k8s.v0.sdcore_config.SdcoreConfigRequires.webui_url",
-        new_callable=PropertyMock,
-    )
-    patcher_is_resource_created = patch(
-        "charms.data_platform_libs.v0.data_interfaces.DatabaseRequires.is_resource_created"
-    )
-    patcher_get_assigned_certificate = patch(
-        "charms.tls_certificates_interface.v4.tls_certificates.TLSCertificatesRequiresV4.get_assigned_certificate"
-    )
-    patcher_db_fetch_relation_data = patch(
-        "charms.data_platform_libs.v0.data_interfaces.DatabaseRequires.fetch_relation_data"
-    )
-
-    @pytest.fixture(autouse=True)
-    def context(self):
-        self.ctx = scenario.Context(
-            charm_type=AMFOperatorCharm,
-        )
-
-    @pytest.fixture(autouse=True)
-    def setup(self, request):
-        self.mock_get_assigned_certificate = (
-            TestCharmCollectUnitStatus.patcher_get_assigned_certificate.start()
-        )
-        self.mock_is_resource_created = (
-            TestCharmCollectUnitStatus.patcher_is_resource_created.start()
-        )
-        self.mock_nrf_url = TestCharmCollectUnitStatus.patcher_nrf_url.start()
-        self.mock_webui_url = TestCharmCollectUnitStatus.patcher_webui_url.start()
-        self.mock_check_output = TestCharmCollectUnitStatus.patcher_check_output.start()
-        self.mock_k8s_service = TestCharmCollectUnitStatus.patcher_k8s_service.start().return_value
-        self.mock_db_fetch_relation_data = (
-            TestCharmCollectUnitStatus.patcher_db_fetch_relation_data.start()
-        )
-        yield
-        request.addfinalizer(self.teardown)
-
-    @staticmethod
-    def teardown() -> None:
-        patch.stopall()
-
+class TestCharmCollectUnitStatus(AMFUnitTestFixtures):
     def test_given_fiveg_nrf_relation_not_created_when_collect_unit_status_then_status_is_blocked(
         self,
     ):
@@ -278,7 +224,7 @@ class TestCharmCollectUnitStatus:
         )
         self.mock_is_resource_created.return_value = True
         self.mock_webui_url.return_value = ""
-        self.mock_nrf_url.return_value = NRF_URL
+        self.mock_nrf_url.return_value = "http://nrf"
 
         state_out = self.ctx.run("collect_unit_status", state_in)
 
@@ -310,7 +256,7 @@ class TestCharmCollectUnitStatus:
             ],
         )
         self.mock_is_resource_created.return_value = True
-        self.mock_nrf_url.return_value = NRF_URL
+        self.mock_nrf_url.return_value = "http://nrf"
 
         state_out = self.ctx.run("collect_unit_status", state_in)
 
@@ -351,7 +297,7 @@ class TestCharmCollectUnitStatus:
             self.mock_get_assigned_certificate.return_value = None, None
             self.mock_check_output.return_value = b"1.1.1.1"
             self.mock_is_resource_created.return_value = True
-            self.mock_nrf_url.return_value = NRF_URL
+            self.mock_nrf_url.return_value = "http://nrf"
 
             state_out = self.ctx.run("collect_unit_status", state_in)
 
@@ -426,7 +372,7 @@ class TestCharmCollectUnitStatus:
             self.mock_get_assigned_certificate.return_value = provider_certificate, private_key
             self.mock_check_output.return_value = b"1.1.1.1"
             self.mock_is_resource_created.return_value = True
-            self.mock_nrf_url.return_value = NRF_URL
+            self.mock_nrf_url.return_value = "http://nrf"
 
             state_out = self.ctx.run("collect_unit_status", state_in)
 
@@ -469,7 +415,7 @@ class TestCharmCollectUnitStatus:
                 ],
             )
             self.mock_check_output.return_value = b""
-            self.mock_nrf_url.return_value = NRF_URL
+            self.mock_nrf_url.return_value = "http://nrf"
 
             state_out = self.ctx.run("collect_unit_status", state_in)
 
@@ -597,7 +543,7 @@ class TestCharmCollectUnitStatus:
             self.mock_k8s_service.get_hostname.return_value = None
             self.mock_k8s_service.get_ip.return_value = None
             self.mock_is_resource_created.return_value = True
-            self.mock_nrf_url.return_value = NRF_URL
+            self.mock_nrf_url.return_value = "http://nrf"
 
             state_out = self.ctx.run("collect_unit_status", state_in)
