+++ tests/unit/test_charm_fiveg_n2_relation_joined.py
@@ -1,51 +1,14 @@
 # Copyright 2024 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-from unittest.mock import PropertyMock, patch
 
-import pytest
 import scenario
 from ops.pebble import Layer, ServiceStatus
 
-from charm import AMFOperatorCharm
-from k8s_service import K8sService
-
-NRF_URL = "http://nrf:8081"
+from tests.unit.fixtures import AMFUnitTestFixtures
 
 
-class TestCharmFiveGN2RelationJoined:
-    patcher_check_output = patch("charm.check_output")
-    patcher_k8s_service = patch("charm.K8sService", autospec=K8sService)
-    patcher_nrf_url = patch(
-        "charms.sdcore_nrf_k8s.v0.fiveg_nrf.NRFRequires.nrf_url", new_callable=PropertyMock
-    )
-    patcher_is_resource_created = patch(
-        "charms.data_platform_libs.v0.data_interfaces.DatabaseRequires.is_resource_created"
-    )
-
-    @pytest.fixture(autouse=True)
-    def context(self):
-        self.ctx = scenario.Context(
-            charm_type=AMFOperatorCharm,
-        )
-
-    @pytest.fixture(autouse=True)
-    def setup(self, request):
-        self.mock_is_resource_created = (
-            TestCharmFiveGN2RelationJoined.patcher_is_resource_created.start()
-        )
-        self.mock_nrf_url = TestCharmFiveGN2RelationJoined.patcher_nrf_url.start()
-        self.mock_check_output = TestCharmFiveGN2RelationJoined.patcher_check_output.start()
-        self.mock_k8s_service = (
-            TestCharmFiveGN2RelationJoined.patcher_k8s_service.start().return_value
-        )
-        yield
-        request.addfinalizer(self.teardown)
-
-    @staticmethod
-    def teardown() -> None:
-        patch.stopall()
-
+class TestCharmFiveGN2RelationJoined(AMFUnitTestFixtures):
     def test_given_service_not_running_when_fiveg_n2_relation_joined_then_n2_information_is_not_in_relation_databag(  # noqa: E501
         self,
     ):
@@ -210,7 +173,7 @@ class TestCharmFiveGN2RelationJoined:
         self.mock_k8s_service.get_hostname.return_value = None
         self.mock_k8s_service.get_ip.return_value = "1.1.1.1"
         self.mock_is_resource_created.return_value = True
-        self.mock_nrf_url.return_value = NRF_URL
+        self.mock_nrf_url.return_value = "http://nrf:8081"
 
         state_out = self.ctx.run(fiveg_n2_relation.joined_event, state_in)
